services/market-data/requirements.txt
fastapi>=0.115,<1
uvicorn>=0.34,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.30,<1
aiokafka>=0.12,<1
alembic>=1.14,<2
httpx>=0.28,<1
python-dateutil>=2.9,<3
